requirements.txt
django==4.0.2
pillow==9.0.1
django-imagekit==4.1.0